Reorganize matching in new file
